requirements.txt
django>=5.2.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
mistralai>=1.7.0
chromadb>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
huey>=2.5.0
redis==6.1.0
langchain>=0.1.0
crawl4ai>=0.6.3
django-filter==23.3
psycopg2-binary==2.9.9
celery==5.3.4
gunicorn==21.2.0